Move the version check to pytorch_mixtral.py
