+++ superbench/benchmarks/model_benchmarks/__init__.py
@@ -13,13 +13,9 @@ from superbench.benchmarks.model_benchmarks.pytorch_cnn import PytorchCNN
 from superbench.benchmarks.model_benchmarks.pytorch_lstm import PytorchLSTM
 from superbench.benchmarks.model_benchmarks.megatron_gpt3 import MegatronGPT
 from superbench.benchmarks.model_benchmarks.pytorch_llama import PytorchLlama
+from superbench.benchmarks.model_benchmarks.pytorch_mixtral import PytorchMixtral
 
-# Check for Python version > 3.7 and conditionally import PytorchMixtral
-PytorchMixtral: Optional[type] = None
-if sys.version_info >= (3, 8):
-    from superbench.benchmarks.model_benchmarks.pytorch_mixtral import PytorchMixtral
-
-__all__ = ['ModelBenchmark', 'PytorchBERT', 'PytorchGPT2', 'PytorchCNN', 'PytorchLSTM', 'MegatronGPT', 'PytorchLlama']
+__all__ = ['ModelBenchmark', 'PytorchBERT', 'PytorchGPT2', 'PytorchCNN', 'PytorchLSTM', 'MegatronGPT', 'PytorchLlama', 'PytorchMixtral']
 
 if PytorchMixtral is not None:
     __all__.append('PytorchMixtral')
